add os Services que faltavam e mudei alguns tipos de atributos,
Retirei tipos primitivos e coloquei do tipo classe, por exemplo float por Double, para permitir que eu compare um valor deste objeto com null

+++ concesswebapi/src/main/java/com/example/concesswebapi/Model/Entity/Carro.java
@@ -17,5 +17,5 @@ public class Carro extends TipoVeiculo {
 
     private String motorizacao;
     private String transmissao;
-    private float potencia;
+    private Double potencia;
 }
